Moved helper function into seperate header files.
